Software/requirements.txt
pyserial
crcmod